+++ pipeline/exec/baseexec.go
@@ -88,10 +88,13 @@ func (e *BaseExecutor) wasmCall(outputGetter execout.ExecutionOutputGetter) (cal
 				message:    panicErr.Error(),
 				stackTrace: call.ExecutionStack,
 			}
-			return nil, fmt.Errorf("block %d: module %q: %w: %s", clock.Number, e.moduleName, ErrWasmDeterministicExec, errExecutor.Error())
+			return nil, fmt.Errorf("block %d: module %q: general wasm execution panicked: %w: %s", clock.Number, e.moduleName, ErrWasmDeterministicExec, errExecutor.Error())
 		}
 		if err != nil {
-			return nil, fmt.Errorf("block %d: module %q: general wasm execution failed: %v", clock.Number, e.moduleName, err)
+			if err := e.ctx.Err(); err != nil {
+				return nil, fmt.Errorf("block %d: module %q: general wasm execution failed: %w: %s", clock.Number, e.moduleName, err)
+			}
+			return nil, fmt.Errorf("block %d: module %q: general wasm execution failed: %w: %s", clock.Number, e.moduleName, ErrWasmDeterministicExec, err)
 		}
 		if e.instanceCacheEnabled {
 			if err := inst.Cleanup(e.ctx); err != nil {